# Rework normalization layers