--- a/test/layers/basic.jl
+++ b/test/layers/basic.jl
@@ -63,6 +63,11 @@ import Flux: activations
     @test Flux.Diagonal(2)([1 2]) == [1 2; 1 2]
     @test Flux.Diagonal(2)([1,2]) == [1,2]
     @test Flux.Diagonal(2)([1 2; 3 4]) == [1 2; 3 4]
+
+    @test Flux.Diagonal(2)(rand(2,3,4)) |> size == (2, 3, 4)
+    @test Flux.Diagonal(2,3)(rand(2,3,4)) |> size == (2, 3, 4)
+    @test Flux.Diagonal(2,3,4)(rand(2,3,4)) |> size == (2, 3, 4)
+    @test Flux.Diagonal(2,3)(rand(2,1,4)) |> size == (2, 3, 4)
   end
 
   @testset "Maxout" begin