This goes in the direction of creating a powerful and consistent experience for the normalization layers. 
- add an activation function to LayerNorm, consistently with other normalization layers
- generalize LayerNorm and Diagonal to handle multiple dimensions
- add the bool keyword `affine` to LayerNorm, GroupNorm, InstanceNorm, and BatchNorm to activate/deactivate a learnable shift and scaling
- add the keyword `track_stats` to InstanceNorm and BatchNorm to activate/deactivate running time mean and variance tracking, to be used in the evaluation phase.
- use pytorch's defaults for `track_stats`  and `affine`, since AFAIK they correspond to common usage and best practice:
   - For InstanceNorm, `affine=false` and `track_stats=false`. This breaks the current behaviour (both `true`)
   - For BatchNorm, `affine=true` and `track_stats=true`, as we are currently doing
   - For GroupNorm, `affine=true` and `track_stats=false` 
- unifies InstanceNorm, GroupNorm, and BatchNorm logics
- perf improvement for InstanceNorm that was previously using repetition instead of broadcasting
- Fix InstanceNorm previously not working on gpu (Fix #1195)
- Fix GroupNorm previously not working on gpu (Fix #1247)
- Fix #1280, fix #1308, fix #1429, fix #1430, fix #802

### PR Checklist

- [x] Tests are added
- [x] Entry in NEWS.md
- [x] Documentation, if applicable
- [ ] Final review from `@dhairyagandhi96` (for API changes).