@CarloLucibello 2020-12-19
> In general it's alright to have extra steps available for different norm layers but not too have to check everything everytime for every layer I think. Otherwise, we end up with one large function which should be multiple smaller functions which are combined appropriately for the layers

Every layer has `affine`, `track_stats`, and `isactive` options. While the `isactive` check has to be performed at runtime, we could maybe rely on type parametrization and dispatch for `affine` and `track_stats` (if this is what you mean, and assuming we don't want to change e.g. track_stats at some point in the execution). Honestly, I don't think removing a couple of `if` is worth the extra complexity involved, nor I believe there would be any performance benefit.   Can you mock up the alternative so I can benchmark it?