@johnnychen94 2020-12-19
> fix InstanceNorm previously not working on gpu

This will fix #1195 if I get it right?