@CarloLucibello 2020-12-24
> This will fix #1195 if I get it right?

yes, this is tested in test/cuda/layers.jl now that InstanceNorm in not on the list of broken layers