@CarloLucibello 2020-12-25
I've also updated GroupNorm and I've done an extensive search of the issues fixed by this PR