@CarloLucibello 2020-12-30
> I'm still a bit uncomfortable with having affine switches in the layers, do we have a better way around this? Perhaps the users are better suited to add that as a layer in the model

I don't see how to do it in a way that doesn't involve a branch somewhere. And we really don't want the user to have to add manually an Affine layer after BatchNorm

Pytorch has the same interface, it can't be that bad if we do the same.