@CarloLucibello 2020-12-30
I did some benchmarking using the script below.
The script was executed 3 times on the current PR
and another 3 times on similar code where I removed the track_stat and affine branching
(so I only keep the path where both are true). 

Fortunately, there is no appreciable performance difference for any system size. 

SCRIPT:
```
using Flux
using BenchmarkTools
using Random 

function test_BN(shape)
    Random.seed!(17)
    x = rand(Float32, shape)
    bn = BatchNorm(shape[end-1])
    @btime $bn($x)
end

for shape in [  (10,10),
                (100,100),
                (1000,1000),]
    test_BN(shape)
end
```
OUTPUT¨:
```julia
# WITHOUT track_stats and affine branching

  4.427 μs (41 allocations: 2.19 KiB)
  38.264 μs (42 allocations: 40.84 KiB)
  3.552 ms (42 allocations: 3.82 MiB)

  4.402 μs (41 allocations: 2.19 KiB)
  38.231 μs (42 allocations: 40.84 KiB)
  3.336 ms (42 allocations: 3.82 MiB)

  4.407 μs (41 allocations: 2.19 KiB)
  37.572 μs (42 allocations: 40.84 KiB)
  3.476 ms (42 allocations: 3.82 MiB)


# WITH track_stats and affine branching

  4.538 μs (41 allocations: 2.19 KiB)
  38.196 μs (42 allocations: 40.84 KiB)
  3.354 ms (42 allocations: 3.82 MiB)

  4.472 μs (41 allocations: 2.19 KiB)
  40.386 μs (42 allocations: 40.84 KiB)
  3.549 ms (42 allocations: 3.82 MiB)

  4.357 μs (41 allocations: 2.19 KiB)
  38.216 μs (42 allocations: 40.84 KiB)
  3.580 ms (42 allocations: 3.82 MiB)
```