@DhairyaLGandhi 2021-01-05
Could you also check the backwards pass?

Having an affine layer seems fine to me, but I'm open to thoughts around construction

As in other test cases on the gpu, I'd much prefer if we didn't have a separate function to go to. Using it to test layers is an extra function to track errors into. We should have the tests be very clear about what they are testing, and best to have that be clear at the site where the test is performed.