@CarloLucibello 2021-01-06
I tested the pullback as well. The results is the following:
- master, this PR, and this PR w/o branching have same performance for `n=1000`
- master's pullback is the slowest for `n=10` and `n=100`
- removing branching we gain ~8% at `n=10` and ~3% at `n=100` in the pullback compared to this PR
- master, PR, and PR w/o branching's pullbacks do a very large number of allocations, something we should investigate (in the future)

My opinion is that
```julia
using Flux
using BenchmarkTools
using Random 
using Zygote: pullback

function test_BN(shape)
    Random.seed!(17)
    println("# Shape $shape")
    x = rand(Float32, shape)
    bn = BatchNorm(shape[end-1])
    # println("### forward")
    # @btime $bn($x)
    println("forward in gradient context")
    @btime pullback(x -> sum(sin.($bn(x))), $x)
    y, back = pullback(x -> sum(sin.(bn(x))), x)
    println("pullback")
    @btime $back(1f0)
end

for shape in [  (10,10),
                (100,100),
                (1000,1000),]
    test_BN(shape)
end
```
```julia
##### Master 

# Shape (10, 10)
forward gradient context
  35.807 μs (508 allocations: 32.97 KiB)
pullback
  250.860 μs (714 allocations: 23.14 KiB)
# Shape (100, 100)
forward gradient context
  207.208 μs (877 allocations: 431.64 KiB)
pullback
  715.362 μs (30697 allocations: 999.34 KiB)
# Shape (1000, 1000)
forward gradient context
  19.116 ms (5457 allocations: 38.32 MiB)
pullback
  54.512 ms (3003397 allocations: 95.48 MiB)

##### This PR

 # Shape (10, 10)
forward in gradient context
  27.025 μs (400 allocations: 29.41 KiB)
pullback
  216.878 μs (708 allocations: 23.17 KiB)
# Shape (100, 100)
forward in gradient context
  192.794 μs (409 allocations: 419.22 KiB)
pullback
  681.668 μs (30692 allocations: 1.01 MiB)
# Shape (1000, 1000)
forward in gradient context
  19.810 ms (409 allocations: 38.21 MiB)
pullback
  53.283 ms (3003392 allocations: 99.29 MiB)

### this PR w/o affine and track_stats branching

# Shape (10, 10)
forward in gradient context
  27.070 μs (365 allocations: 27.00 KiB)
pullback
  198.775 μs (693 allocations: 22.67 KiB)
# Shape (100, 100)
forward in gradient context
  194.250 μs (374 allocations: 416.81 KiB)
pullback
  661.045 μs (30677 allocations: 1.01 MiB)
# Shape (1000, 1000)
forward in gradient context
  20.032 ms (374 allocations: 38.20 MiB)
pullback
  54.227 ms (3003377 allocations: 99.29 MiB)
```