@DhairyaLGandhi 2021-01-06
I don't think zygote is much of a culprit here. Besides, we should add an adjoint for norm, and that would make it much faster