@CarloLucibello 2021-01-08
> I don't think zygote is much of a culprit here. Besides, we should add an adjoint for norm, and that would make it much faster

What do you mean by an adjoint for `norm`? The whole `_norm_layer_forward` function? That seems quite painful to do